backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pandas>=2.0
numpy>=2.0
scikit-learn>=1.5
xgboost>=3.2.0
python-dotenv==1.1.1
pytest==8.4.1
httpx==0.28.1
statsmodels>=0.14.0
torch>=2.0.0
shap>=0.46.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
groq>=0.11.0
pypdf>=4.3.0
reportlab>=4.2.0
